refactor(modulos): F3 cirúrgica PR-1 — routePlanner é telefonia (−45 arestas, 16% da baseline)
O par telefonia↔reposicao (45 arestas) era erro de corte, não acoplamento:
components/reposicao/routePlanner/ tinha todos os consumidores (AdminRoutePlanner,
useRoutePlanner, lib/route/*) e todas as dependências na telefonia, e zero uso
pela reposição. git mv → components/rota/planner/ (24 arquivos, histórico
preservado) + import-fix em 17 consumidores; manifesto intacto (o glob
components/rota/** da telefonia já capturava o destino — gate F1 confirma);
baseline de fronteiras re-gerada 284→239 (−45 líquido, 6 re-rotuladas).

Prova: gates 67/67 · tsc verde · lint 0 erros · suíte 4746/4746.

Re-decisão do PASSO 0 da F3 (founder + Codex gpt-5.5 convergentes) registrada
em docs/historico/modularizacao.md: move completo descartado → 3 PRs cirúrgicos.
