+++ docs/historico/modularizacao.md
@@ -2,6 +2,13 @@
 
 > Programa de separação do app em módulos avaliáveis (bugs/qualidade/velocidade por módulo). Specs em `docs/superpowers/specs/2026-07-08-modularizacao-f*`. Ordem executada (decisão do founder 2026-07-08, acatando parecer Codex): **F1→F2→F4**, com **F3 (mover código físico) adiada** para janela coordenada dedicada — e re-decidir lá se ainda vale.
 
+## 🔪 F3 — re-decidida (2026-07-09): move completo DESCARTADO → programa cirúrgico em 3 PRs
+
+- **Re-decisão (founder + Codex `gpt-5.5` reasoning high):** o parecer de 2026-07-08 se confirmou e reforçou — *"o problema restante é fronteira lógica, não diretório; mover tudo para `src/modules/` não reduz uma única aresta, só troca conflito de arquitetura por conflito de Git"*. Contexto que pesou: PR #1212 (~100 arquivos de src/) parado + 8 sessões paralelas vivas. Founder aprovou o cirúrgico guiado pelo relatório de fronteiras.
+- **Critério Codex para burn-down (vale para os próximos):** aresta **legítima** = capacidade estável expressa como verbo, sem importar UI/hook/contexto interno do outro módulo; **vazamento** = importa interno alheio, cria bidirecionalidade ou fan-out. Par bidirecional com dezenas de arestas = módulo mal cortado ou contrato ausente. ⚠️ Anti-padrão nomeado: plataforma como **"lixão de contratos"** — contrato em plataforma só com linguagem transversal (aparecer `Farmer`/`SPIN`/`RoutePlanner` no nome = acoplamento lavado).
+- **PR-1 — routePlanner → telefonia (este PR).** As 45 arestas do par telefonia↔reposicao (16% da baseline F2) eram **erro de corte, não acoplamento**: `components/reposicao/routePlanner/` tinha todos os consumidores (`AdminRoutePlanner`, `useRoutePlanner`, `lib/route/*`) e todas as dependências (`lib/route`, `lib/maps`, `BotaoLigar`) na telefonia; zero consumo pela reposição. `git mv` → `components/rota/planner/` (24 arquivos, histórico preservado) + import-fix em 17 consumidores. **Manifesto intacto** (glob `components/rota/**` da telefonia já capturava o destino — gate F1 confirmou sem órfão/sobreposição). Baseline **284→239** (−45 líquido; 6 arestas re-rotuladas honestamente p/ farmer↔telefonia, alvo do PR-2). Método que achou o alvo: detalhar a baseline por par (`grep deModulo/paraModulo`) — o fan-out 100% concentrado numa pasta denunciou o corte errado.
+- **Próximos:** PR-2 contrato SPIN (11 das 17 arestas telefonia→farmer; quebra o ciclo — tem design a fazer, cuidado com o anti-padrão acima); PR-3 sino do AppShell (7 inversões plataforma→negócio + ~30 KB eager, já quantificado na F4).
+
 ## ✅ F2 — gate de fronteiras anti-vazamento com ratchet (2026-07-08, PR #1255)
 
 - **Regra**: negócio importa só de si+plataforma; plataforma↛negócio exceto `COMPOSICAO_RAIZ` (`App.tsx`); `import type` conta (kind no diagnóstico); teste de A importando B = acoplamento A→B.
