+++ src/components/rota/planner/ManualModeCard.tsx
