+++ src/components/rota/planner/StatsStrip.tsx
