+++ src/lib/modulos/fronteiras-baseline.ts
@@ -87,22 +87,12 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/components/OrderSuccessDialog.tsx", para: "src/lib/reposicao.ts", deModulo: "loja-afiacao", paraModulo: "reposicao", kind: "runtime" },
   { de: "src/components/OrderSuccessDialog.tsx", para: "src/services/orderSubmission/index.ts", deModulo: "loja-afiacao", paraModulo: "vendas", kind: "type" },
   { de: "src/components/radar/RadarDetailSheet.tsx", para: "src/components/call/BotaoLigar.tsx", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/AlvosFiltros.tsx", para: "src/lib/route/field-targets.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
-  { de: "src/components/reposicao/routePlanner/AlvosFiltros.tsx", para: "src/lib/route/prospect-stop.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/CityMultiSelector.tsx", para: "src/lib/route/city-filter.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/CityMultiSelector.tsx", para: "src/queries/useRadarCidadesRota.ts", deModulo: "reposicao", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/FieldTargetCard.tsx", para: "src/components/call/BotaoLigar.tsx", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/FieldTargetDetailSheet.tsx", para: "src/components/call/BotaoLigar.tsx", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/FieldTargetDetailSheet.tsx", para: "src/lib/route/alvo-detalhe.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
-  { de: "src/components/reposicao/routePlanner/MapLegend.tsx", para: "src/lib/route/marker-visual.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/RouteStopCard.tsx", para: "src/components/call/BotaoLigar.tsx", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/RouteStopCard.tsx", para: "src/components/radar/RadarOutcomeMenu.tsx", deModulo: "reposicao", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/ScheduledVisitsPanel.tsx", para: "src/hooks/useVisitasAgendadas.ts", deModulo: "reposicao", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/ScheduledVisitsPanel.tsx", para: "src/lib/maps/nav-link.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/ScheduledVisitsPanel.tsx", para: "src/lib/visitas/today.ts", deModulo: "reposicao", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/ScheduledVisitsPanel.tsx", para: "src/lib/visitas/visita-status.ts", deModulo: "reposicao", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/reposicao/routePlanner/UfSelector.tsx", para: "src/lib/route/city-filter.ts", deModulo: "reposicao", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
   { de: "src/components/RequireFinanceiroAccess.tsx", para: "src/services/financeiroV2Service.ts", deModulo: "plataforma", paraModulo: "financeiro", kind: "runtime" },
+  { de: "src/components/rota/planner/CityMultiSelector.tsx", para: "src/queries/useRadarCidadesRota.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
+  { de: "src/components/rota/planner/RouteStopCard.tsx", para: "src/components/radar/RadarOutcomeMenu.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
+  { de: "src/components/rota/planner/ScheduledVisitsPanel.tsx", para: "src/hooks/useVisitasAgendadas.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
+  { de: "src/components/rota/planner/ScheduledVisitsPanel.tsx", para: "src/lib/visitas/today.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
+  { de: "src/components/rota/planner/ScheduledVisitsPanel.tsx", para: "src/lib/visitas/visita-status.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/components/sales/print/buildPrintHtml.ts", para: "src/components/OrderPrintLayout.tsx", deModulo: "vendas", paraModulo: "loja-afiacao", kind: "runtime" },
   { de: "src/components/salesOrders/print.ts", para: "src/components/OrderPrintLayout.tsx", deModulo: "vendas", paraModulo: "loja-afiacao", kind: "runtime" },
   { de: "src/components/salesOrders/useSalesOrders.ts", para: "src/utils/whatsappShare.ts", deModulo: "vendas", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
@@ -167,9 +157,6 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/hooks/usePersona.ts", para: "src/lib/dashboard/persona-detect.ts", deModulo: "plataforma", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/usePersona.ts", para: "src/lib/dashboard/route-tracker.ts", deModulo: "plataforma", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/useProcessComparison.ts", para: "src/lib/customer-process/comparison-types.ts", deModulo: "knowledge-base", paraModulo: "admin-crm", kind: "type" },
-  { de: "src/hooks/useRoutePlanner.ts", para: "src/components/reposicao/routePlanner/constants.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/hooks/useRoutePlanner.ts", para: "src/components/reposicao/routePlanner/priority.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/hooks/useRoutePlanner.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
   { de: "src/hooks/useRoutePlanner.ts", para: "src/hooks/useFarmerScoring.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/useRoutePlanner.ts", para: "src/lib/visitas/agenda-to-stop.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/useSpinAnalysis.ts", para: "src/lib/transcription/types.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
@@ -195,20 +182,7 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/lib/offline-handlers.ts", para: "src/services/recebimento-confirm.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
   { de: "src/lib/offline-handlers.ts", para: "src/services/recebimento-cte.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
   { de: "src/lib/offline-handlers.ts", para: "src/services/recebimento-divergencia.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
-  { de: "src/lib/route/alvo-detalhe.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/alvo-detalhe.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
   { de: "src/lib/route/alvo-detalhe.ts", para: "src/lib/radar/ui-helpers.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/lib/route/cidades-cache.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/cidades-cache.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/city-filter.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/city-filter.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/field-targets.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/field-targets.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/geocode-fila.cep.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/geocode-fila.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/geocode-fila.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/marker-visual.test.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
-  { de: "src/lib/route/marker-visual.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
   { de: "src/lib/standard-process/types.ts", para: "src/lib/customer-process/types.ts", deModulo: "knowledge-base", paraModulo: "admin-crm", kind: "type" },
   { de: "src/lib/tint/preflight-formulas.ts", para: "src/lib/preco/parse-decimal-br.ts", deModulo: "tintometrico", paraModulo: "vendas", kind: "runtime" },
   { de: "src/lib/venda-assistida/__tests__/selos.test.ts", para: "src/lib/knowledge-base/catalisador-link.ts", deModulo: "vendas", paraModulo: "knowledge-base", kind: "runtime" },
@@ -225,25 +199,6 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/pages/AdminOrderDetail.tsx", para: "src/components/OrderChat.tsx", deModulo: "admin-crm", paraModulo: "loja-afiacao", kind: "runtime" },
   { de: "src/pages/AdminOrderDetail.tsx", para: "src/components/SendingQualityChecklist.tsx", deModulo: "admin-crm", paraModulo: "loja-afiacao", kind: "runtime" },
   { de: "src/pages/AdminReposicaoCockpit.tsx", para: "src/components/dataHealth/DataHealthBanner.tsx", deModulo: "reposicao", paraModulo: "governanca", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/AlvosFiltros.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/CheckoutDialog.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/CityMultiSelector.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/constants.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/FieldTargetDetailSheet.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/FieldTargetsList.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/FieldTargetsSummary.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/ManualModeCard.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/MapLegend.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/PeriodFilter.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/PlanningModeSelector.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/renderHelpers.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/RouteActionButtons.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/RoutePlannerContextTabs.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/RouteStopCard.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/ScheduledVisitsPanel.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/StatsStrip.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/TodayVisitCard.tsx", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "runtime" },
-  { de: "src/pages/AdminRoutePlanner.tsx", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "reposicao", kind: "type" },
   { de: "src/pages/Customer360.tsx", para: "src/components/tarefas/VozTarefaDialog.tsx", deModulo: "farmer-inteligencia", paraModulo: "tarefas", kind: "runtime" },
   { de: "src/pages/Customer360.tsx", para: "src/hooks/useCustomerContacts.ts", deModulo: "farmer-inteligencia", paraModulo: "admin-crm", kind: "runtime" },
   { de: "src/pages/Customer360.tsx", para: "src/hooks/useReguaPreco360.ts", deModulo: "farmer-inteligencia", paraModulo: "vendas", kind: "runtime" },
@@ -284,7 +239,7 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/pages/UnifiedOrder.tsx", para: "src/utils/whatsappShare.ts", deModulo: "vendas", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
   { de: "src/pages/VendasFerramentas.tsx", para: "src/pages/FarmerBundles.tsx", deModulo: "vendas", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/pages/VendasFerramentas.tsx", para: "src/pages/FarmerRecommendations.tsx", deModulo: "vendas", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/queries/useRadarCidadesRota.ts", para: "src/components/reposicao/routePlanner/types.ts", deModulo: "farmer-inteligencia", paraModulo: "reposicao", kind: "type" },
+  { de: "src/queries/useRadarCidadesRota.ts", para: "src/components/rota/planner/types.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/queries/useRadarCidadesRota.ts", para: "src/lib/route/cidades-cache.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
   { de: "src/queries/useRegistrarContato.ts", para: "src/lib/route/route-outcome.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/services/orderSubmission/buildPrintData.ts", para: "src/components/OrderPrintLayout.tsx", deModulo: "vendas", paraModulo: "loja-afiacao", kind: "type" },
