+++ src/pages/AdminRoutePlanner.tsx
@@ -5,31 +5,31 @@ import { Loader2, Navigation, CheckCircle2 } from 'lucide-react';
 import { PageSkeleton } from '@/components/ui/page-skeleton';
 import L from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-import { formatDuration } from '@/components/reposicao/routePlanner/renderHelpers';
-import { STOP_CONFIG } from '@/components/reposicao/routePlanner/constants';
-import { StatsStrip } from '@/components/reposicao/routePlanner/StatsStrip';
-import { RouteStopCard } from '@/components/reposicao/routePlanner/RouteStopCard';
-import { TodayVisitCard } from '@/components/reposicao/routePlanner/TodayVisitCard';
-import { CheckoutDialog } from '@/components/reposicao/routePlanner/CheckoutDialog';
-import { PlanningModeSelector } from '@/components/reposicao/routePlanner/PlanningModeSelector';
-import { RoutePlannerContextTabs } from '@/components/reposicao/routePlanner/RoutePlannerContextTabs';
-import { CityMultiSelector } from '@/components/reposicao/routePlanner/CityMultiSelector';
-import { FieldTargetsSummary } from '@/components/reposicao/routePlanner/FieldTargetsSummary';
-import { AlvosFiltros } from '@/components/reposicao/routePlanner/AlvosFiltros';
-import { FieldTargetsList } from '@/components/reposicao/routePlanner/FieldTargetsList';
-import { FieldTargetDetailSheet } from '@/components/reposicao/routePlanner/FieldTargetDetailSheet';
-import type { RouteStop } from '@/components/reposicao/routePlanner/types';
-import { PeriodFilter } from '@/components/reposicao/routePlanner/PeriodFilter';
-import { RouteActionButtons } from '@/components/reposicao/routePlanner/RouteActionButtons';
-import { ManualModeCard } from '@/components/reposicao/routePlanner/ManualModeCard';
-import { ScheduledVisitsPanel } from '@/components/reposicao/routePlanner/ScheduledVisitsPanel';
+import { formatDuration } from '@/components/rota/planner/renderHelpers';
+import { STOP_CONFIG } from '@/components/rota/planner/constants';
+import { StatsStrip } from '@/components/rota/planner/StatsStrip';
+import { RouteStopCard } from '@/components/rota/planner/RouteStopCard';
+import { TodayVisitCard } from '@/components/rota/planner/TodayVisitCard';
+import { CheckoutDialog } from '@/components/rota/planner/CheckoutDialog';
+import { PlanningModeSelector } from '@/components/rota/planner/PlanningModeSelector';
+import { RoutePlannerContextTabs } from '@/components/rota/planner/RoutePlannerContextTabs';
+import { CityMultiSelector } from '@/components/rota/planner/CityMultiSelector';
+import { FieldTargetsSummary } from '@/components/rota/planner/FieldTargetsSummary';
+import { AlvosFiltros } from '@/components/rota/planner/AlvosFiltros';
+import { FieldTargetsList } from '@/components/rota/planner/FieldTargetsList';
+import { FieldTargetDetailSheet } from '@/components/rota/planner/FieldTargetDetailSheet';
+import type { RouteStop } from '@/components/rota/planner/types';
+import { PeriodFilter } from '@/components/rota/planner/PeriodFilter';
+import { RouteActionButtons } from '@/components/rota/planner/RouteActionButtons';
+import { ManualModeCard } from '@/components/rota/planner/ManualModeCard';
+import { ScheduledVisitsPanel } from '@/components/rota/planner/ScheduledVisitsPanel';
 import { useRoutePlanner } from '@/hooks/useRoutePlanner';
 import { escapeHtml } from '@/lib/escape-html';
 import 'leaflet.markercluster';
 import 'leaflet.markercluster/dist/MarkerCluster.css';
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css';
 import { markerVisual, clusterStats, precisaoVisual, TONE_CSS, type MarkerTone, type MarkerShape } from '@/lib/route/marker-visual';
-import { MapLegend } from '@/components/reposicao/routePlanner/MapLegend';
+import { MapLegend } from '@/components/rota/planner/MapLegend';
 
 // Fix default marker icons for Leaflet + bundlers
 delete (L.Icon.Default.prototype as { _getIconUrl?: unknown })._getIconUrl;
